+++ packages/react-dev-utils/package.json
@@ -19,7 +19,8 @@
     "openBrowser.js",
     "prompt.js",
     "WatchMissingNodeModulesPlugin.js",
-    "webpackHotDevClient.js"
+    "webpackHotDevClient.js",
+    "webpackEnvDefinePlugin.js"
   ],
   "dependencies": {
     "ansi-html": "0.0.5",
@@ -28,6 +29,7 @@
     "html-entities": "1.2.0",
     "opn": "4.0.2",
     "sockjs-client": "1.0.3",
-    "strip-ansi": "3.0.1"
+    "strip-ansi": "3.0.1",
+    "webpack": "^1.13.3"
   }
 }
