+++ packages/react-dev-utils/webpackEnvDefinePlugin.js
