+++ packages/react-scripts/package.json
@@ -64,7 +64,6 @@
     "url-loader": "0.5.7",
     "webpack": "1.13.2",
     "webpack-dev-server": "1.16.2",
-    "webpack-env-define-plugin": "1.0.0",
     "webpack-manifest-plugin": "1.1.0",
     "whatwg-fetch": "1.0.0"
   },
