--- packages/webpack-env-define-plugin/README.md
@@ -1,46 +1,0 @@
-# webpack-env-define-plugin
-
-This package includes a webpack plugin used by [Create React App](https://github.com/facebookincubator/create-react-app).  
-The plugin grabs environment variables that follows a specified regex and injects them into the application using the 
-webpack DefinePlugin.
-In addition, the plugin allows defining custom environment variables to be injected.
-
-## Usage in Create React App Projects
-
-See [webpack.config.dev.js](https://github.com/facebookincubator/create-react-app/blob/master/packages/react-scripts/config/webpack.config.dev.js#L200) and [webpack.config.prod.js](https://github.com/facebookincubator/create-react-app/blob/master/packages/react-scripts/config/webpack.config.prod.js#L234).
-
-## Usage Outside of Create React App
-
-If you want to use this webpack plugin in a project not built with Create React App, you can install it with following steps.
-
-First, install this package.
-
-  ```
-  npm install --save-dev webpack-env-define-plugin
-  ```
-
-Then reference it in your webpack config:
-
-  ```js
-  var EnvDefinePlugin = require('../../webpack-env-define-plugin');
-  ```
-
-and define it as a plugin in the webpack config plugin section:
- 
- ```js
-  plugins: [
-     ...
-     new EnvDefinePlugin({
-       // Grab MY_PREFIX_* environment variables
-       regex: /^MY_PREFIX_/i,
-       customVariables: {
-         // Useful for determining whether we’re running in production mode.
-         // Most importantly, it switches React into the correct mode.
-         'MY_ENV_VAR': JSON.stringify(
-           process.env.MY_ENV_VAR || 'default'
-         )
-       }
-     }),
-     ...
-  ]
- ```
